Splitted with_siblings
