+++ app/models/concerns/on_chapter.rb
@@ -2,7 +2,7 @@ module OnChapter
   extend ActiveSupport::Concern
 
   included do
-    include WithSiblings
+    include WithParent
 
     has_one :chapter_guide
     has_one :chapter, through: :chapter_guide
